+++ test/issues/include_module.mli
@@ -10,7 +10,12 @@ end
 type nonrec t = t
 
 (* {gospel_expected|
-   [125] File "include_module.mli", line 10, characters 16-17:
+   [125] File "include_module.mli", line 6, characters 0-24:
+         6 | include sig
+         7 |   type t
+         8 | end
+         Warning: `include`s are currently ignored by the gospel type-checker.
+         File "include_module.mli", line 10, characters 16-17:
          10 | type nonrec t = t
                               ^
          Error: Symbol t not found in scope