+++ CHANGES.md
@@ -25,6 +25,8 @@
 
 ## Improved
 
+- Display a warning when encountering an `include`
+  [\#334](https://github.com/ocaml-gospel/gospel/pull/334)
 - Allow patterns in arguments and return type annotation in anonymous functions
   [\#309](https://github.com/ocaml-gospel/gospel/pull/309)
 - Propagate pattern locations to report errors to the precise patterns